requirements.txt
about-time==4.2.1
alive-progress==3.2.0
asttokens==3.0.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.2.1
einops==0.8.1
exceptiongroup==1.2.2
executing==2.1.0
feedparser==6.0.11
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
grapheme==0.6.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.34.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.4.2
numpy==1.23.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pickleshare==0.7.5
pillow==11.1.0
pip==25.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==5.9.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==308
PyYAML==6.0.2
pyzmq==24.0.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
setuptools==75.8.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.6
stack_data==0.6.3
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tornado==6.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.1
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uv==0.6.10
wcwidth==0.2.13
wheel==0.45.1
zipp==3.21.0
